pyproject.toml
[build-system]
requires = ["setuptools", "setuptools_scm[toml]"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages=[
    "gammapy_tools", 
    "gammapy_tools.make_background", 
    "gammapy_tools.fake_source_coordinates", 
    "gammapy_tools.templates",
    "gammapy_tools.analysis",
    "gammapy_tools.utils",
    ]

[tool.setuptools_scm]
write_to = "gammapy_tools/_version.py"

[project]
name = "gammapy_tools"
dynamic = ["version"]
description = "VERITAS (VEGAS and Eventdisplay) to DL3 Converter"
readme = "README.md"
license = { file="LICENSE" }

requires-python = ">=3.9"
#dependencies = [
#    "gammapy==1.2",
#    "pydantic<=1.10",
#    "scipy==1.11.4",
#    "astropy==5.3.4",
#    "astroquery==0.4.6",
#    "matplotlib==3.8.2",
#    "multiprocess==0.70.16",
#    "numpy==1.26.4",
#    #"pyV2DL3==0.5",
#    "PyYAML==6.0.1",
#    "regions==0.8",
#]

dependencies = [
    "gammapy>=1.1",
    "pydantic",
    "scipy==1.11.4",
    "astropy",
    "astroquery",
    "matplotlib",
    "multiprocess",
    "numpy",
    #"pyV2DL3",
    "PyYAML",
    "regions",
]

[project.optional-dependencies]
"tests" = [
    "pytest",
    "pytest-cov",
]

[project.urls]
#"documentation" = "https://github.com/VERITAS-Observatory/gammapy-tools/"
#"repository" = "https://github.com/VERITAS-Observatory/gammapy-tools/"
#"bug tracker" = "https://github.com/VERITAS-Observatory/gammapy-tools/issues"

[project.scripts]

[tool.commitizen]
version = "0.1.3"  # This should be your current semver version